+++ IdeaProjects/mimimips/src/UFRRJ/arquitetura/laboratorio/Memoria.java
@@ -7,9 +7,9 @@ import java.util.ArrayList;
  */
 public class Memoria {
 
-    public static int[] memValor = new int[128];//memoria que irá guardar os storewords etc
+    public static long[] memValor = new long[128];//memoria que irá guardar os storewords etc
 
-    public static ArrayList<Integer> memWord = new ArrayList<Integer>();//memoria que irá guadar as words, array list pq vai ter que ser em sequencia msm, então meh
+    public static ArrayList<Long> memWord = new ArrayList<Long>();//memoria que irá guadar as words, array list pq vai ter que ser em sequencia msm, então meh
     //todo lógica do stackpointer
